feat: sync feat/jexpanel-sync — initial working version
- Fixed vite.config.ts: removed manualChunks to fix Temporal Dead Zone (TDZ) errors
- Fixed non-null assertion crashes in AuthRouter, Login/Register containers
- Fixed IntelligenceController -> AI/SettingsController namespace
- Added ApplicationApiController alias for missing namespace
- Added all JexPanel-extracted API modules and components
- Added billing, legal, tickets, webhooks, presets, passkeys, SSH key features
- Added activity logging, AI settings, theme designer
- Fixed MySQL migrations for FK/index compat
- Version v1.0.0.0

+++ database/migrations/2026_04_12_150000_create_auto_scaling_rules_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('auto_scaling_rules', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('server_id');
+            $table->unsignedInteger('server_id');
             
             // Thresholds
             $table->unsignedTinyInteger('cpu_threshold')->default(80)->comment('CPU threshold in percent');
